refactor: getParentCommitFromCommitHash 함수 commitUtil에서 resetUtil로 위치 이동
