+++ src/main/kotlin/geet/utils/commandutil/porcelainutil/commitUtil.kt
@@ -3,8 +3,6 @@ package geet.utils.commandutil.porcelainutil
 import geet.exceptions.NotFound
 import geet.utils.GEET_DIR_PATH
 import geet.utils.GEET_HEAD_FILE_PATH
-import geet.utils.GEET_OBJECTS_DIR_PATH
-import geet.utils.compressToZlib
 import java.io.File
 
 fun getCurrentRef(): String {
@@ -32,18 +30,4 @@ fun editCurrentRefContent(content: String) {
     }
 
     refFile.writeText(content)
-}
-
-fun getParentCommitFromCommitHash(commitHash: String): String {
-    val dirName = commitHash.substring(0, 2)
-    val fileName = commitHash.substring(2)
-    val commitFile = File("${GEET_OBJECTS_DIR_PATH}/${dirName}/${fileName}")
-    if (!commitFile.exists()) {
-        throw NotFound("커밋 파일을 찾을 수 없습니다.")
-    }
-
-    val commitContents = commitFile.readText()
-    val commitContentsDecompressed = compressToZlib(commitContents)
-    val commitContentsSplit = commitContentsDecompressed.split("\n")
-    return commitContentsSplit[1].split(" ")[1]
 }
