+++ src/main/kotlin/geet/utils/commandutil/porcelainutil/resetUtil.kt
@@ -1,7 +1,9 @@
 package geet.utils.commandutil.porcelainutil
 
 import geet.exceptions.BadRequest
+import geet.exceptions.NotFound
 import geet.utils.GEET_OBJECTS_DIR_PATH
+import geet.utils.decompressFromZlib
 import java.io.File
 
 fun changeToFullHash(commitString: String): String {
@@ -46,4 +48,18 @@ fun countCarrot(string: String): Int {
 fun isHash(hash: String): Boolean {
     val pattern = Regex("^[0-9a-f]{4,40}$")
     return hash.matches(pattern)
+}
+
+fun getParentCommitFromCommitHash(commitHash: String): String {
+    val dirName = commitHash.substring(0, 2)
+    val fileName = commitHash.substring(2)
+    val commitFile = File("${GEET_OBJECTS_DIR_PATH}/${dirName}/${fileName}")
+    if (!commitFile.exists()) {
+        throw NotFound("커밋 파일을 찾을 수 없습니다.")
+    }
+
+    val commitContents = commitFile.readText()
+    val commitContentsDecompressed = decompressFromZlib(commitContents)
+    val commitContentsSplit = commitContentsDecompressed.split("\n")
+    return commitContentsSplit[1].split(" ")[1]
 }
